runWithDistCache as private test helper.
